Make buildah's "--pull-always" a bit more optional
In some situations in CI, we want to use a FROM image that's from a previous
build in that same CI chain.

If we have:

	cat > first/Dockerfile < EOF
	FROM scratch
	EOF

	cat > second/Dockerfile < EOF
	second/Dockerfile
	FROM first:latest
	EOF

Then, for this CI pipeline, we first want to build "first",  load this
container and tag it _locally_ on the matchine as `first:latest`, and
only after that perform the build on `second`.

Then, once the two builds have both gone through CI and tests, do we want to
publish them to the registry under these names.

By adding an option to allow us to override the `--pull-always` flag in those
situations, we can change it to `--pull-never`, which permits us to solve this
problem.