Add the option of binding a directory as a volume during build
For artifacts that should be available, but not permanently part of the image,
it's sometimes better to not package it up in a tarball and add it as a layer,
but rather to just ship it in as a volume.

For example, if a container is built from a set of packages, and not a source
check out, it's better to add the files as a volume during build.

Volumes are added as read-only, as we do not want them to be changed from
inside the build.

The envisioned usecase is to pass in a directory full of python wheel files,
and then install them in the container.