Generate podman/buildah compatible source archives
With this change the SOURCE_ARCHIVE goal archives the commit tree
object HEAD^{tree} rather than the commit object HEAD. The reason for
this is to circumvent a problem with podman/buildah which will
currently error out if it finds a global extended header when adding a
tar archive.

The git archive command will generate tar archives with a global
extended header containing the commit hash as a comment if the
archived object is a commit.