Make "temp-publish" behave like "publish"
Podman 3.2.0-rc has changes to it's docker:// transport layer and now
raises an error if publishing a container using only "docker://" syntax
for remote url.

This is because it tries to look up the image in the docker:// transport
as well, something that does not work.

This change makes the two publish steps work the same.

Issue: ModioAB/agile#2252