Allow custom Dockerfile ARGs for container builds
This allows a container build to pass any Dockerfile `ARG`s it needs 
into the container environment while building, e.g. to pass software 
versions, when running the process inside the container, rather than 
importing a pre-baked tarball.

Previously we had a special case for the `FROM` image, and this 
generalizes that to a list of `VAR[=value]` that gets mapped as 
`--build-arg=VAR[=value]`.